LPS-205477 Remove scr pojo service wiring design in BookmarksFolderBatchReindexer
