+++ modules/apps/bookmarks/bookmarks-service/src/main/java/com/liferay/bookmarks/internal/search/BookmarksEntryModelSearchConfigurator.java
@@ -8,9 +8,12 @@ package com.liferay.bookmarks.internal.search;
 import com.liferay.bookmarks.internal.search.spi.model.index.contributor.BookmarksEntryModelIndexerWriterContributor;
 import com.liferay.bookmarks.internal.search.spi.model.result.contributor.BookmarksEntryModelSummaryContributor;
 import com.liferay.bookmarks.model.BookmarksEntry;
+import com.liferay.bookmarks.model.BookmarksFolder;
 import com.liferay.bookmarks.service.BookmarksEntryLocalService;
 import com.liferay.portal.kernel.search.Field;
 import com.liferay.portal.search.batch.DynamicQueryBatchIndexingActionableFactory;
+import com.liferay.portal.search.indexer.IndexerDocumentBuilder;
+import com.liferay.portal.search.indexer.IndexerWriter;
 import com.liferay.portal.search.spi.model.index.contributor.ModelIndexerWriterContributor;
 import com.liferay.portal.search.spi.model.registrar.ModelSearchConfigurator;
 import com.liferay.portal.search.spi.model.result.contributor.ModelSummaryContributor;
@@ -56,7 +59,9 @@ public class BookmarksEntryModelSearchConfigurator
 	protected void activate() {
 		_modelIndexWriterContributor =
 			new BookmarksEntryModelIndexerWriterContributor(
-				_bookmarksEntryLocalService, _bookmarksFolderBatchReindexer,
+				_bookmarksEntryLocalService,
+				new BookmarksFolderBatchReindexer(
+					_indexerDocumentBuilder, _indexerWriter),
 				_dynamicQueryBatchIndexingActionableFactory);
 	}
 
@@ -64,11 +69,18 @@ public class BookmarksEntryModelSearchConfigurator
 	private BookmarksEntryLocalService _bookmarksEntryLocalService;
 
 	@Reference
-	private BookmarksFolderBatchReindexer _bookmarksFolderBatchReindexer;
-
-	@Reference
 	private DynamicQueryBatchIndexingActionableFactory
 		_dynamicQueryBatchIndexingActionableFactory;
+
+	@Reference(
+		target = "(indexer.class.name=com.liferay.bookmarks.model.BookmarksFolder)"
+	)
+	private IndexerDocumentBuilder _indexerDocumentBuilder;
+
+	@Reference(
+		target = "(indexer.class.name=com.liferay.bookmarks.model.BookmarksFolder)"
+	)
+	private IndexerWriter<BookmarksFolder> _indexerWriter;
 
 	private ModelIndexerWriterContributor<BookmarksEntry>
 		_modelIndexWriterContributor;
