--- modules/apps/bookmarks/bookmarks-service/src/main/java/com/liferay/bookmarks/internal/search/BookmarksFolderBatchReindexerImpl.java
@@ -1,56 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.bookmarks.internal.search;
-
-import com.liferay.bookmarks.model.BookmarksFolder;
-import com.liferay.portal.kernel.dao.orm.Property;
-import com.liferay.portal.kernel.dao.orm.PropertyFactoryUtil;
-import com.liferay.portal.search.batch.BatchIndexingActionable;
-import com.liferay.portal.search.indexer.IndexerDocumentBuilder;
-import com.liferay.portal.search.indexer.IndexerWriter;
-
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
-
-/**
- * @author Luan Maoski
- */
-@Component(service = BookmarksFolderBatchReindexer.class)
-public class BookmarksFolderBatchReindexerImpl
-	implements BookmarksFolderBatchReindexer {
-
-	@Override
-	public void reindex(long folderId, long companyId) {
-		BatchIndexingActionable batchIndexingActionable =
-			indexerWriter.getBatchIndexingActionable();
-
-		batchIndexingActionable.setAddCriteriaMethod(
-			dynamicQuery -> {
-				Property folderIdProperty = PropertyFactoryUtil.forName(
-					"folderId");
-
-				dynamicQuery.add(folderIdProperty.eq(folderId));
-			});
-		batchIndexingActionable.setCompanyId(companyId);
-		batchIndexingActionable.setPerformActionMethod(
-			(BookmarksFolder bookmarksFolder) ->
-				batchIndexingActionable.addDocuments(
-					indexerDocumentBuilder.getDocument(bookmarksFolder)));
-
-		batchIndexingActionable.performActions();
-	}
-
-	@Reference(
-		target = "(indexer.class.name=com.liferay.bookmarks.model.BookmarksFolder)"
-	)
-	protected IndexerDocumentBuilder indexerDocumentBuilder;
-
-	@Reference(
-		target = "(indexer.class.name=com.liferay.bookmarks.model.BookmarksFolder)"
-	)
-	protected IndexerWriter<BookmarksFolder> indexerWriter;
-
-}
